move youkiinstance to its own module
