+++ crates/containerd-shim-wasmedge/src/instance.rs
@@ -7,8 +7,9 @@ use std::sync::{Arc, Condvar, Mutex};
 use anyhow::Context;
 use anyhow::Result;
 use containerd_shim_wasm::sandbox::error::Error;
-use containerd_shim_wasm::sandbox::instance::{ExitCode, YoukiInstance};
+use containerd_shim_wasm::sandbox::instance::ExitCode;
 use containerd_shim_wasm::sandbox::instance_utils::maybe_open_stdio;
+use containerd_shim_wasm::sandbox::youki_instance::YoukiInstance;
 use containerd_shim_wasm::sandbox::{EngineGetter, InstanceConfig};
 use libc::{dup2, STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO};
 use nix::unistd::close;
